+++ frontend/src/app/login/LoginFrom.tsx
@@ -1,0 +1,48 @@
+"use client";
+
+import { useState } from "react";
+import { login } from "@/utils/api";
+import { useRouter } from "next/navigation";
+
+const LoginForm = () => {
+  const router = useRouter();
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    try {
+      const user = await login(email, password);
+      alert(`Welcome back, ${user.username}!`);
+      router.push("/");
+    } catch (error) {
+      alert("Login failed!");
+    }
+  };
+
+  return (
+    <form onSubmit={handleSubmit} className="space-y-4 p-4 border rounded">
+      <input
+        type="email"
+        value={email}
+        onChange={(e) => setEmail(e.target.value)}
+        placeholder="Email"
+        className="border p-2 w-full"
+        required
+      />
+      <input
+        type="password"
+        value={password}
+        onChange={(e) => setPassword(e.target.value)}
+        placeholder="Password"
+        className="border p-2 w-full"
+        required
+      />
+      <button type="submit" className="bg-blue-500 text-white p-2 rounded">
+        Log In
+      </button>
+    </form>
+  );
+};
+
+export default LoginForm;
