+++ frontend/src/app/login/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import LoginForm from "./LoginFrom";
+
+const page = () => {
+  return <LoginForm />;
+};
+
+export default page;
