+++ frontend/src/app/signup/SignupForm.tsx
@@ -1,0 +1,56 @@
+"use client";
+
+import { useState } from "react";
+import { signup } from "@/utils/api";
+import { useRouter } from "next/navigation";
+
+const SignupForm = () => {
+  const router = useRouter();
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [username, setUsername] = useState("");
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    try {
+      await signup(username, email, password);
+      alert("Signup successful! Please log in.");
+      router.push("/login");
+    } catch (error) {
+      alert("Signup failed!");
+    }
+  };
+
+  return (
+    <form onSubmit={handleSubmit} className="space-y-4 p-4 border rounded">
+      <input
+        value={username}
+        onChange={(e) => setUsername(e.target.value)}
+        placeholder="Username"
+        className="border p-2 w-full"
+        required
+      />
+      <input
+        type="email"
+        value={email}
+        onChange={(e) => setEmail(e.target.value)}
+        placeholder="Email"
+        className="border p-2 w-full"
+        required
+      />
+      <input
+        type="password"
+        value={password}
+        onChange={(e) => setPassword(e.target.value)}
+        placeholder="Password"
+        className="border p-2 w-full"
+        required
+      />
+      <button type="submit" className="bg-blue-500 text-white p-2 rounded">
+        Sign Up
+      </button>
+    </form>
+  );
+};
+
+export default SignupForm;
