+++ frontend/src/app/signup/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import AuthForm from "./AuthForm";
+import AuthForm from "./SignupForm";
 
 const page = () => {
   return <AuthForm />;
